src/test/java/com/liv/test/algo/backtrack/WordBreakTest.java
package com.liv.test.algo.backtrack;

import com.liv.algo.backtrack.WordBreak;
import com.liv.algo.dp.WordBreakDp;
import org.junit.Assert;
import org.junit.Test;

import java.util.ArrayList;
import java.util.List;

public class WordBreakTest {

    private static final WordBreak wordBreak = new WordBreak();

    private static final WordBreakDp wordBreakDp = new WordBreakDp();

    @Test
    public void test(){
        String s = "leetcode";
        List<String> wordDict = new ArrayList<>();
        wordDict.add("leet");
        wordDict.add("code");
        wordDict.add("aa");
        Assert.assertEquals(true, wordBreak.wordBreak(s,wordDict));
    }

    @Test
    public void testDp1(){
        String s = "leetcode";
        List<String> wordDict = new ArrayList<>();
        wordDict.add("leet");
        wordDict.add("code");
        wordDict.add("aa");
        Assert.assertEquals(true, wordBreakDp.wordBreak(s,wordDict));
    }

    @Test
    public void test1(){
        String s = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaab";
        List<String> wordDict = new ArrayList<>();
        wordDict.add("a");
        wordDict.add("aa");
        wordDict.add("aaaa");
        wordDict.add("aaaaa");
        wordDict.add("aaaaaaa");
        wordDict.add("aaaaaaaa");
        wordDict.add("aaaaaaaaa");
        wordDict.add("aaaaaaaaaa");
        Assert.assertEquals(false, wordBreak.wordBreak(s,wordDict));
    }

    @Test
    public void testDp2(){
        String s = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaab";
        List<String> wordDict = new ArrayList<>();
        wordDict.add("a");
        wordDict.add("aa");
        wordDict.add("aaaa");
        wordDict.add("aaaaa");
        wordDict.add("aaaaaaa");
        wordDict.add("aaaaaaaa");
        wordDict.add("aaaaaaaaa");
        wordDict.add("aaaaaaaaaa");
        Assert.assertEquals(false, wordBreakDp.wordBreak(s,wordDict));
    }
}
